Moving IFimGetSetPixel to primitives
