+++ src/primitives/IFimGetSetPixel.ts
